Implement ParallelRead()
As part of this change, I'm also removing the CreateSession() function that
served as a simple proof of concept.

ParallelRead() is close to its final shape: it creates a new read session and
returns a vector of ReadSession objects. In a future change, I will add the
ability to pass a ReadSession to a Read() function.